Extract deployment logic to execute.js
